+++ R/differential_expression.R
@@ -910,9 +910,9 @@ FindMarkers.DimReduc <- function(
 #'
 FindMarkers.Seurat <- function(
   object,
-  latent.vars = NULL,
   ident.1 = NULL,
   ident.2 = NULL,
+  latent.vars = NULL,
   group.by = NULL,
   subset.ident = NULL,
   assay = NULL,