+++ tests/testthat/test_differential_expression.R
@@ -251,11 +251,11 @@ test_that("t test works", {
 
 results <- suppressWarnings(
   FindMarkers(
-    object = pbmc_small, 
-    ident.1 = 0, ident.2 = 1, 
-    test.use = "negbinom", 
-    verbose = FALSE, 
-    base = exp(1), 
+    object = pbmc_small,
+    ident.1 = 0, ident.2 = 1,
+    test.use = "negbinom",
+    verbose = FALSE,
+    base = exp(1),
     fc.slot = "counts",
     pseudocount.use = 1
   )
@@ -272,12 +272,12 @@ test_that("negbinom test works", {
 
 results <- suppressWarnings(
   FindMarkers(
-    object = pbmc_small, 
-    ident.1 = 0, 
-    ident.2 = 1, 
-    test.use = "poisson", 
-    verbose = FALSE, 
-    base = exp(1), 
+    object = pbmc_small,
+    ident.1 = 0,
+    ident.2 = 1,
+    test.use = "poisson",
+    verbose = FALSE,
+    base = exp(1),
     fc.slot = "counts",
     pseudocount.use = 1
   )
@@ -430,7 +430,7 @@ query.list <- lapply(X = query.list, FUN = ScaleData, verbose = FALSE)
 query.list <- suppressWarnings(lapply(X = query.list, FUN = RunPCA, verbose = FALSE, npcs = 20))
 
 anchors <- suppressMessages(suppressWarnings(FindIntegrationAnchors(object.list = c(ref, query.list), k.filter = NA, verbose = FALSE)))
-object <- suppressMessages(IntegrateData(anchorset = anchors,  k.weight = 25, verbose = FALSE))
+object <- suppressWarnings(suppressMessages(IntegrateData(anchorset = anchors,  k.weight = 25, verbose = FALSE)))
 object <- suppressMessages(ScaleData(object, verbose = FALSE))
 object <- suppressMessages(RunPCA(object, verbose = FALSE))
 object <- suppressMessages(FindNeighbors(object = object, verbose = FALSE))