+++ NEWS.md
@@ -1,12 +1,16 @@
 # Unreleased
 
 ## Changes
-- Added `VisiumV2` class - inherits `SeuratObject::FOV`
+- Updated interactive `SpatialDimPlot`s to display spatial coordinates
+- Updated `BuildClusterTree` to `PseudobulkExpression(..., method = "aggregate")` instead of `AverageExpression`
+- Fixed `SpatialPlot` to properly handle images with shared cells
+- Added `cluster.name` parameter to `BuildNicheAssay`
+- Added `VisiumV2` class, inherits `SeuratObject::FOV`, returned by `Load10X_Spatial`
 - Updated `Load10X_Spatial` to support loading Visium HD datasets - added `bin.size` parameter 
 - Updated `Read10X_Coordinates` to handle new `tissue_coordinates.parquet` files
 - Added `shape` parameter to `SpatialPlot` and wrappers: `SpatialFeaturePlot` and `SpatialDimPlot`
-- Added `image.scale` parameter to `SpatialPlot` and wrappers: `SpatialFeaturePlot` and `SpatialDimPlot`
-- Added `image.scale` parameter to `Read10X_Image`
+- Added `image.scale` parameter to `SpatialPlot` and related functions: `SpatialFeaturePlot`/`ISpatialFeaturePlot` and `SpatialDimPlot`/`ISpatialDimPlot`
+- Restored `image.name` parameter to `Read10X_Image`
 - Added `scale` parameter to `Radius.VisiumV1`
 - Fixed calculation of `spot.radius` attribute for `VisiumV1` instance returned by `Read10X_Image` — replace `scale.factors$fiducial` with `scale.factors$spot`
 - Added `Read10X_Coordinates` and `Read10X_ScaleFactors`